metadata: use jwt.ParseWithClaims instead of mapstructure
Just use the jwt library to pick apart the claims object rather
than introducing a separate unmarshaling step from a map to an
object. Also, increase the width of one of the metadata entry
integer fields, because the production metadata won't parse
without it. (I believe this is a bug exposed by dropping the
old mapstructure library.)